+++ Products/CMFPlone/interfaces/controlpanel.py
@@ -1024,8 +1024,15 @@ class INavigationSchema(Interface):
             u"The content types that should be shown in the navigation and "
             u"site map."),
         required=False,
-        default=('Image', 'File', 'Link', 'News Item', 'Folder', 'Document',
-                 'Event'),
+        default=(
+            'Image',
+            'File',
+            'Link',
+            'News Item',
+            'Folder',
+            'Document',
+            'Event'
+        ),
         value_type=schema.Choice(
             source="plone.app.vocabularies.ReallyUserFriendlyTypes"
         ))